backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy>=1.26.0
onnxruntime>=1.17.0
mediapipe>=0.10.13
pydantic==2.5.0
opencv-python-headless==4.8.1.78
huggingface-hub>=0.20.0  # For downloading models from Hugging Face

# Test dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests>=2.31.0  # For test_deployed_api.py script

